requirements.txt
tensorflow-cpu==2.8.0
scikit-image==0.19.2
Pillow==8.2.0
numpy==1.20
matplotlib==3.4.2
Flask==2.0.1
protobuf==3.17.3
# opencv-python==4.5.3.56
Flask-Cors==3.0.10